examples/requirements.txt
# HarborAI Examples 依赖包列表

# 核心依赖
harborai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# 异步支持
asyncio>=3.4.3
aiohttp>=3.8.0
aiofiles>=23.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
json5>=0.9.0
scipy>=1.11.0
scikit-learn>=1.3.0

# 日志和监控
structlog>=23.0.0
prometheus-client>=0.17.0
psutil>=5.9.0

# 数据库支持 (可选)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.11.0
asyncpg>=0.28.0

# 测试和开发
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# 可视化 (用于某些示例)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
streamlit>=1.25.0

# 企业级功能
redis>=4.6.0
celery>=5.3.0
flower>=2.0.0

# 安全库
cryptography>=41.0.0
bcrypt>=4.0.0

# 系统监控
psutil>=5.9.0

# JSON Schema 验证
jsonschema>=4.17.0
PyJWT>=2.8.0

# 配置管理
pyyaml>=6.0.0
toml>=0.10.0

# HTTP 客户端
httpx>=0.24.0
requests>=2.31.0
certifi>=2023.7.22

# 时间处理
python-dateutil>=2.8.0
pytz>=2023.3

# 文本处理
jinja2>=3.1.0
markdown>=3.4.0

# 开发工具
rich>=13.0.0
click>=8.1.0
tqdm>=4.65.0